separate out pack_summmary; test_vis passes (2/3 tries)
